% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviceUtils.R
\name{getTikzDefaults}
\alias{getTikzDefaults}
\title{Get default options set by tikzDevice.}
\usage{
getTikzDefaults()
}
\value{
This function returns a list with elements
\itemize{
\item \code{tikzDefaultEngine}
\item \code{tikzLatex}
\item \code{tikzDocumentDeclaration}
\item \code{tikzFooter}
\item \code{tikzLatexPackages}
\item \code{tikzXelatexPackages}
\item \code{tikzLualatexPackages}
\item \code{tikzMetricPackages}
\item \code{tikzUnicodeMetricPackages}
\item \code{tikzSanitizeCharacters}
\item \code{tikzReplacementCharacters}
\item \code{tikzPdftexWarnUTF}
}
}
\description{
Get default options set by tikzDevice.
}
