% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviceUtils.R
\name{tikzTest}
\alias{tikzTest}
\title{Test invocation of a LaTeX engine.}
\usage{
tikzTest(engine = getOption("tikzDefaultEngine"),
  documentDeclaration = getOption("tikzDocumentDeclaration"), packages,
  verbose = TRUE)
}
\arguments{
\item{engine}{a string specifying which TeX engine to use. Possible values
are 'pdftex', 'xetex' and 'luatex'. See the Unicode section of
\link{tikzDevice-package} for details.}

\item{documentDeclaration}{See the sections ``Options That Affect Package
Behavior'' and ``Font Size Calculations'' of \link{tikzDevice-package}
for more details.}

\item{packages}{See the section ``Options That Affect Package Behavior'' of
\link{tikzDevice-package}.}

\item{verbose}{If set to \code{FALSE}, calling this function will not cause any output to
be printed to the screen. Defaults to \code{TRUE}.}
}
\description{
This function simulates the measurement of dimensions and prints detailed
information in case of errors.
}
\seealso{
\code{\link{tikz}}
}
