% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviceUtils.R
\name{tikzTest}
\alias{tikzTest}
\title{Test invocation of a LaTeX engine.}
\usage{
tikzTest(
  texString = "A",
  engine = getOption("tikzDefaultEngine"),
  documentDeclaration = getOption("tikzDocumentDeclaration"),
  packages
)
}
\arguments{
\item{texString}{An arbitrary string for which the width is to be
calculated.  May contain LaTeX markup.}

\item{engine}{a string specifying which TeX engine to use. Possible values
are 'pdftex', 'xetex' and 'luatex'. See the Unicode section of
\link{tikzDevice-package} for details.}

\item{documentDeclaration}{See the sections "Options That Affect Package
Behavior" and "Font Size Calculations" of \link{tikzDevice-package}
for more details.}

\item{packages}{See the section "Options That Affect Package Behavior" of
\link{tikzDevice-package}.}
}
\description{
This function simulates the measurement of dimensions and prints detailed
information in case of errors.
}
\seealso{
\code{\link[=tikz]{tikz()}}
}
